Move the table list headers into the component.
